Update mount() tests
